+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Rules/Utxo.hs
@@ -54,7 +54,7 @@ import Cardano.Ledger.Alonzo.Rules.Utxos (
   UtxosEnv (..),
  )
 import Cardano.Ledger.Alonzo.Scripts (OrdExUnits (..), pointWiseExUnits)
-import Cardano.Ledger.Alonzo.Tx (AlonzoEraTx (..), IsPhase2Valid (..), totExUnits)
+import Cardano.Ledger.Alonzo.Tx (AlonzoEraTx (..), IsPhase2Valid (..))
 import Cardano.Ledger.Alonzo.TxBody (
   AllegraEraTxBody (..),
   AlonzoEraTxBody (..),
@@ -462,10 +462,7 @@ validateWrongNetworkInTxBody netId txBody =
 --
 -- > totExunits tx ≤ maxTxExUnits pp
 validateExUnitsTooBigUTxO ::
-  ( AlonzoEraTxWits era
-  , EraTx era
-  , AlonzoEraPParams era
-  ) =>
+  AlonzoEraTx era =>
   PParams era ->
   Tx l era ->
   Test (AlonzoUtxoPredFailure era)
@@ -476,7 +473,7 @@ validateExUnitsTooBigUTxO pp tx =
   where
     maxTxExUnits = pp ^. ppMaxTxExUnitsL
     -- This sums up the ExUnits for all embedded Plutus Scripts anywhere in the transaction:
-    totalExUnits = totExUnits tx
+    totalExUnits = getTotalExUnits tx
 
 -- | Ensure that number of collaterals does not exceed the allowed @maxCollInputs@ parameter.
 --